rename week to plan
